Remove raw string input when parsing docstring as a string.
1. When a docstring is parsed from a python file (by reading the python
file as a string and not as a module), the back slashes within raw string,
r""", needs to be converted to the appropriate format. However since the
docstring is read in as a string, all occurences of the backslash is
converted to escape sequences when succeeded by the right character.
This string cannot just be turned into the raw form (through repr)
because existing escape sequences are turned into their impotent
counterpart.