Fix bug in utils.wrap.
If subsequent indents and initial indents are different, then the
maximum number of characters for each line is different. This is
problematic when the subsequent indent is calculated from some stating
line. For example,

    x = ('123'
         '456')

can be created by first making an initial 'x = (' and wrapping around
the text '123456'. If special precautions are not taken, the first line
can exceed the maxium width, while the subsequent lines do not.

Some keywords were introduced to fix this, and tests were added.

The API needs to be changed because it is not too messy.