Remove trailing newline when extracting math equations
Need different behaviour when math equation is in different sections.
When it is in extended summary section, trailing newline is not
necessary, but it is necessary when it is in other sections. It is
easier to assume that there are no trailing newlines than it is to check
if there exists trailing newlines and adding it when necessary.